parse int and uint

+++ cli/spam.go
@@ -48,7 +48,7 @@ func (h *Handler) BuildSpammer(sh throughput.SpamHelper, defaults bool) (*throug
 		return throughput.NewSpammer(sc, sh)
 	}
 	// Collect parameters
-	numAccounts, err := prompt.Int("number of accounts", consts.MaxInt)
+	numAccounts, err := prompt.Int("number of accounts", int64(consts.MaxInt))
 	if err != nil {
 		return nil, err
 	}
@@ -64,19 +64,19 @@ func (h *Handler) BuildSpammer(sh throughput.SpamHelper, defaults bool) (*throug
 		return nil, err
 	}
 
-	txsPerSecond, err := prompt.Int("txs to try and issue per second", consts.MaxInt)
+	txsPerSecond, err := prompt.Int("txs to try and issue per second", int64(consts.MaxInt))
 	if err != nil {
 		return nil, err
 	}
-	minTxsPerSecond, err := prompt.Int("minimum txs to issue per second", consts.MaxInt)
+	minTxsPerSecond, err := prompt.Int("minimum txs to issue per second", int64(consts.MaxInt))
 	if err != nil {
 		return nil, err
 	}
-	txsPerSecondStep, err := prompt.Int("txs to increase per second", consts.MaxInt)
+	txsPerSecondStep, err := prompt.Int("txs to increase per second", int64(consts.MaxInt))
 	if err != nil {
 		return nil, err
 	}
-	numClients, err := prompt.Int("number of clients per node", consts.MaxInt)
+	numClients, err := prompt.Int("number of clients per node", int64(consts.MaxInt))
 	if err != nil {
 		return nil, err
 	}
@@ -86,11 +86,11 @@ func (h *Handler) BuildSpammer(sh throughput.SpamHelper, defaults bool) (*throug
 		key,
 		sZipf,
 		vZipf,
-		txsPerSecond,
-		minTxsPerSecond,
-		txsPerSecondStep,
-		numClients,
-		numAccounts,
+		int(txsPerSecond),
+		int(minTxsPerSecond),
+		int(txsPerSecondStep),
+		int(numClients),
+		int(numAccounts),
 	)
 
 	return throughput.NewSpammer(sc, sh)
